+++ tests/graphics/geom.js
@@ -2,3 +2,4 @@
 
 import './geom/Dimension.js' ;
 import './geom/Point.js' ;
+import './geom/Rectangle.js' ;